# src/docs/swagger.yaml
openapi: 3.0.3
info:
    title: Fetcher API
    version: 1.0.0
    description: |
        REST API for discovering and executing fetchers and managing their persisted results.

        Protected operations accept JWT bearer tokens issued for this API. Depending on the
        operation, the token must represent either a user with the documented minimum system
        role or an authenticated service.
servers:
    - url: /
...
paths:
...
    /api/v1/fetchers:
        get:
            security:
                - UserBearerAuth: []
            tags:
                - Fetchers
            summary: Get all fetchers
            operationId: getFetchers
            description: |
                Retrieves all available fetchers and their metadata.
                Requires a user token with the `ADMIN` or `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetchers retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/Fetcher'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/fetchers/{fetcherId}:
        parameters:
            - $ref: '#/components/parameters/fetcherId'
        get:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetchers
            summary: Get fetcher by ID
            operationId: getFetcherById
            description: Retrieves a specific fetcher. Accepts either a user token or a service token.
            responses:
                '200':
                    description: Fetcher retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/Fetcher'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/fetchers/{fetcherId}/fetch:
        parameters:
            - $ref: '#/components/parameters/fetcherId'
        post:
            security:
                - UserBearerAuth: []
            tags:
                - Fetchers
            summary: Execute a fetcher
            operationId: fetchFetcher
            description: |
                Executes the selected fetcher directly, without reading or writing `FetchResult`
                cache entries. This endpoint is intended for manual or diagnostic executions.

                Temporal rules:

                | Fetcher capability | `CAPTURE` | `REPLAY` |
                | --- | --- | --- |
                | `SNAPSHOT` | Calls the external source for `effectiveAt`. | Rejected with `400`; a snapshot source cannot reconstruct an uncaptured past instant. |
                | `HISTORICAL` | Calls the external source for `effectiveAt`. | Calls the historical external source for `effectiveAt`. |

                Requires a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/FetchFetcherBody'
            responses:
                '200':
                    description: Fetcher result generated successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetcherExecutionResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
    /api/v1/fetchers/{fetcherId}/validate:
        parameters:
            - $ref: '#/components/parameters/fetcherIdCandidate'
        post:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetchers
            summary: Validate fetcher configuration
            operationId: validateFetcher
            description: |
                Checks whether the fetcher exists and, when `fetcherConfig` is supplied, validates
                it against the fetcher's schema. An unknown `fetcherId` is reported in the successful
                response as `valid: false`.

                Accepts a service token or a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/ValidateFetcherBody'
            responses:
                '200':
                    description: Fetcher configuration validation completed (valid or invalid).
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/ValidateFetcherResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ──────────────────────────────────────────────
    # Fetch Results
    # ──────────────────────────────────────────────
...
    /api/v1/fetchers/{fetcherId}/fetchResults/{fetchResultId}:
        parameters:
            - $ref: '#/components/parameters/fetcherId'
            - $ref: '#/components/parameters/fetchResultId'
        get:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Get fetch result by fetcher ID and fetch result ID
            operationId: getFetchResultByFetcherIdAndFetchResultId
            description: |
                Retrieves a single fetch result by fetcher ID and fetch result ID.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetch result retrieved successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        put:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Update fetch result by fetcher ID and fetch result ID
            operationId: updateFetchResultByFetcherIdAndFetchResultId
            description: |
                Partially updates a fetch result by fetcher ID and fetch result ID.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            requestBody:
                $ref: '#/components/requestBodies/UpdateFetchResultBody'
            responses:
                '200':
                    description: Fetch result updated successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
        delete:
            security:
                - UserBearerAuth: []
                - ServiceBearerAuth: []
            tags:
                - Fetch Results
            summary: Delete fetch result by fetcher ID and fetch result ID
            operationId: deleteFetchResultByFetcherIdAndFetchResultId
            description: |
                Deletes a fetch result by fetcher ID and fetch result ID.
                Accepts a service token or a user token with the `SUPERADMIN` system role.
            responses:
                '200':
                    description: Fetch result deleted successfully.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessResponse'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/FetchResult'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/InternalError'
components:
    securitySchemes:
        UserBearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: |
                JWT whose `type` claim is `user`. Operations also state the minimum accepted
                `systemRole` (`ADMIN` or `SUPERADMIN`). Role priority is
                `USER < ADMIN < SUPERADMIN`.
        ServiceBearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: JWT whose `type` claim is `service`.

    # ── Parameters ────────────────────────────
    parameters:
        fetcherId:
            name: fetcherId
            in: path
            required: true
            description: |
                Registered fetcher identifier. Current examples:
                `FT_REST_BLUEJAY_REPORTER_LOGS`, `FT_GQL_GITHUB_PROJECTV2_ITEMS`,
                `FT_GQL_GITHUB_PULL_REQUESTS`, `FT_GQL_GITHUB_ISSUES`, and
                `FT_GQL_ZENHUB_ISSUES`.
            schema:
                type: string
                example: FT_GQL_GITHUB_ISSUES
        fetchResultId:
            name: fetchResultId
            in: path
            required: true
            description: MongoDB ObjectId of the fetch result
            schema:
                type: string
                pattern: '^[a-fA-F0-9]{24}$'
                example: 67c9ff7b2ddf1b3940f9f321
        fetcherIdCandidate:
            name: fetcherId
            in: path
            required: true
            description: Fetcher identifier to check; it does not have to be registered.
            schema:
                type: string
                example: FT_GQL_GITHUB_ISSUES
...
    requestBodies:
        FetchFetcherBody:
            description: |
                Request body for a direct, non-persisted execution. The fetcher's registered
                `temporalCapability` determines whether the requested mode is supported.
            required: true
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            temporalContext:
                                $ref: '#/components/schemas/TemporalContext'
                            fetcherConfig:
                                $ref: '#/components/schemas/FetcherConfig'
                        required:
                            - temporalContext
                            - fetcherConfig
                    examples:
                        snapshotCapture:
                            summary: Direct observation of a snapshot source
                            value:
                                temporalContext:
                                    effectiveAt: '2026-04-10T09:30:00.000Z'
                                    mode: CAPTURE
                                fetcherConfig:
                                    owner: governify
                                    repository: fetcher
                                    token: github_pat_example
                        historicalReplay:
                            summary: Direct historical query for a past instant
                            value:
                                temporalContext:
                                    effectiveAt: '2026-04-01T09:30:00.000Z'
                                    mode: REPLAY
                                fetcherConfig:
                                    latitude: 37.39
                                    longitude: -5.99
        ValidateFetcherBody:
            description: Optional request body containing the configuration to validate
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            fetcherConfig:
                                $ref: '#/components/schemas/FetcherConfig'
                    example:
                        fetcherConfig:
                            owner: governify
                            repository: fetcher
                            token: github_pat_example
...
        UpdateFetchResultBody:
            description: Request body for updating a fetch result
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/UpdateFetchResultRequest'
                    example:
                        status: COMPLETED
                        startDate: '2026-04-06T10:59:55.000Z'
                        effectiveAt: '2026-04-06T11:00:00.000Z'
                        endDate: '2026-04-06T11:00:00.000Z'
                        data: This is a placeholder for GitHub fetcher data

    # ── Reusable responses ────────────────────────
    responses:
        Unauthorized:
            description: Missing, malformed, invalid, expired, or unsupported token
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Authorization header missing or malformed
                        httpStatus: 401
                        appCode: UNAUTHORIZED
                        data: null
                        error:
                            message: Authorization header missing or malformed
        Forbidden:
            description: Authenticated user does not have the required system role
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Insufficient permissions
                        httpStatus: 403
                        appCode: FORBIDDEN
                        data: null
                        error:
                            message: Insufficient permissions
        ValidationError:
            description: Validation error in the submitted data
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Validation failed
                        httpStatus: 400
                        appCode: VALIDATION_ERROR
                        data: null
                        error:
                            message: Validation failed
                            details:
                                - type: field
                                  value: null
                                  msg: fetcherConfig is required
                                  path: fetcherConfig
                                  location: body
        NotFound:
            description: The requested resource does not exist
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Resource not found
                        httpStatus: 404
                        appCode: NOT_FOUND
                        data: null
                        error:
                            message: Resource not found
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    example:
                        success: false
                        message: Internal server error
                        httpStatus: 500
                        appCode: UNKNOWN_ERROR
                        data: null
                        error:
                            message: Internal server error
    schemas:
...
        JsonValue:
            description: Any JSON-compatible value, including null
            anyOf:
                - type: object
                  nullable: true
                  additionalProperties: true
                - type: array
                  items: {}
                - type: string
                - type: number
                - type: boolean
        SuccessResponse:
            type: object
            description: Standard wrapper for all successful API responses
            additionalProperties: false
            properties:
                success:
                    type: boolean
                    example: true
                message:
                    type: string
                    example: OK
                httpStatus:
                    type: integer
                    example: 200
                appCode:
                    type: string
                    example: SUCCESS
                data:
                    $ref: '#/components/schemas/JsonValue'
                error:
                    type: object
                    nullable: true
                    example: null
            required:
                - success
                - message
                - httpStatus
                - appCode
                - data
                - error
...
        Fetcher:
            type: object
            description: Serialized fetcher definition and metadata; the executable `fetch` function is omitted from JSON.
            additionalProperties: false
            properties:
                id:
                    type: string
                    description: Registered fetcher identifier
                    example: FT_GQL_GITHUB_ISSUES
                temporalCapability:
                    type: string
                    enum: [SNAPSHOT, HISTORICAL]
                    description: |
                        Temporal behavior registered by the fetcher:

                        - `SNAPSHOT`: the external source only exposes its current state. External access
                          is allowed in `CAPTURE`; `REPLAY` can only reuse an already stored result.
                        - `HISTORICAL`: the source can answer for `effectiveAt`, so missing results may be
                          fetched in both `CAPTURE` and `REPLAY`.
                    example: SNAPSHOT
                moreInfo:
                    $ref: '#/components/schemas/FetcherMoreInfo'
                fetcherConfigSchema:
                    type: object
                    description: Internal Zod schema serialized as JSON. Its structure is library-specific.
                    additionalProperties: true
                fetchScript:
                    type: string
                    description: String representation of the fetcher's executable function
                    example: 'async _fetcherConfig => ({ data: [] })'
            required:
                - id
                - temporalCapability
                - moreInfo
                - fetcherConfigSchema
                - fetchScript
        FetcherMoreInfo:
            type: object
            description: Human-readable fetcher metadata
            additionalProperties: false
            properties:
                title:
                    type: string
                    example: GitHub Issues Fetcher
                description:
                    type: string
                    example: Fetches every raw issue of a GitHub repository.
                example:
                    type: string
                    example: ''
            required:
                - title
                - description
                - example
...
        FetcherExecutionResult:
            type: object
            description: Result returned by a fetcher execution
            additionalProperties: false
            properties:
                data:
                    $ref: '#/components/schemas/JsonValue'
            required:
                - data
            example:
                data: []
...
        ValidateFetcherResult:
            type: object
            description: Fetcher existence and configuration validation result
            additionalProperties: false
            properties:
                valid:
                    type: boolean
                    example: true
                error:
                    type: string
                    example: Invalid fetcherConfig
                issues:
                    type: array
                    items:
                        $ref: '#/components/schemas/ValidateFetcherIssue'
            required:
                - valid
        FetchResult:
            type: object
            description: |
                Persisted temporal result. Within a fetcher's collection, `effectiveAt` plus `configHash`
                form the unique logical key. `COMPLETED`, `UNAVAILABLE`, and `IN_PROGRESS` results are
                reused. A `FAILED` result is retried by updating the same document.
            additionalProperties: false
            properties:
                _id:
                    type: string
                    description: Stable document identifier, preserved across retries of a `FAILED` result.
                    example: 67c9ff7b2ddf1b3940f9f321
                startDate:
                    type: string
                    format: date-time
                    description: When the current execution was claimed.
                    example: '2026-04-06T10:59:55.000Z'
                endDate:
                    type: string
                    format: date-time
                    nullable: true
                    description: When the current execution reached a terminal state; null while `IN_PROGRESS`.
                    example: '2026-04-06T11:00:00.000Z'
                effectiveAt:
                    type: string
                    format: date-time
                    description: Exact instant represented by this result and part of its logical cache key.
                    example: '2026-04-06T11:00:00.000Z'
                temporalCapability:
                    type: string
                    enum: [SNAPSHOT, HISTORICAL]
                    description: Capability of the fetcher recorded when this result was created or last retried.
                acquisitionMode:
                    type: string
                    enum: [CAPTURE, REPLAY]
                    description: |
                        Mode that created or last retried this document. On a cache hit this can differ
                        from the current request's mode because the stored result is returned unchanged.
                status:
                    type: string
                    enum: [IN_PROGRESS, COMPLETED, UNAVAILABLE, FAILED]
                    description: |
                        Result lifecycle:

                        - `IN_PROGRESS`: an asynchronous or concurrent external operation is running.
                        - `COMPLETED`: data was obtained successfully and is available in `data`.
                        - `UNAVAILABLE`: the requested instant cannot be resolved under the temporal rules;
                          this is a terminal, reusable domain outcome, not an HTTP error.
                        - `FAILED`: the external operation failed; the next generate request reclaims and retries it.
                    example: COMPLETED
                unavailableReason:
                    type: string
                    nullable: true
                    enum: [PAST_SNAPSHOT_NOT_CAPTURED, HISTORICAL_DATA_NOT_AVAILABLE]
                    description: |
                        Reason for `UNAVAILABLE`; null for every other status.

                        - `PAST_SNAPSHOT_NOT_CAPTURED`: replay requested for a snapshot that was not stored.
                        - `HISTORICAL_DATA_NOT_AVAILABLE`: reserved for a historical source that can
                          explicitly report that the instant has no data. Generic source errors produce `FAILED`.
                fetcherConfig:
                    $ref: '#/components/schemas/FetcherConfig'
                configHash:
                    type: string
                    description: SHA-256 hash of the canonical fetcher configuration and part of the logical cache key.
                    pattern: '^[a-f0-9]{64}$'
                    example: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                data:
                    $ref: '#/components/schemas/JsonValue'
                createdAt:
                    type: string
                    format: date-time
                    example: '2026-04-06T10:59:55.000Z'
                updatedAt:
                    type: string
                    format: date-time
                    example: '2026-04-06T11:00:00.000Z'
                __v:
                    type: integer
                    description: MongoDB document version
                    example: 0
            required:
                - _id
                - startDate
                - endDate
                - effectiveAt
                - temporalCapability
                - acquisitionMode
                - status
                - unavailableReason
                - fetcherConfig
                - configHash
                - data
                - createdAt
                - updatedAt
                - __v
...